feat(plugins): add conform.nvim for formatting refactor(plugins): continue with structuring overhaul
