+++ lua/core/util.lua
@@ -39,6 +39,13 @@ function M.foldtext()
   return ret
 end
 
+function M.formatexpr()
+  if Util.has('conform.nvim') then
+    return require('conform').formatexpr()
+  end
+  return vim.lsp.formatexpr({ timeout_ms = 3000 })
+end
+
 M.skip_foldexpr = {} ---@type table<number,boolean>
 local skip_check = assert(vim.loop.new_check())
